[Refactor] move socket into MVC & wrap socket error handling with curring & closure
